RNG count is standard integer type

+++ model/model.h
@@ -63,7 +63,7 @@ struct state {
 };
 
 struct message {
-  long rng_calls;
+  uint16_t rng_calls;
   enum event_t event;
   struct population people;
 };
@@ -90,7 +90,7 @@ struct population p_normalize(const struct population *,
 struct population p_sample(tw_rng_stream *,
 			   const struct population *,
 			   unsigned int);
-struct population p_exposed(const struct population *, long int *);
+struct population p_exposed(const struct population *, uint16_t *);
 struct population p_person(enum health_t);
 struct population p_right_shift(const struct population *);
 struct population p_left_shift(const struct population *);
@@ -98,6 +98,6 @@ struct population p_left_shift(const struct population *);
 tw_lpid transition_select(tw_rng_stream *,
 			  const struct transition *,
 			  tw_lpid,
-			  long int *);
+			  uint16_t *);
 
 #endif // MODEL_H
